Split RepositoryMethodDefinition into separate files
